Minor: consolidate datafusion_substrait tests into `substrait_integration`
